feat: 옵션 컴포넌트 분리
